% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNAseq_ScoreRegion.R
\name{check_RNAseq_df}
\alias{check_RNAseq_df}
\title{Check Bulk RNA-seq Region data file}
\usage{
check_RNAseq_df(Adf)
}
\arguments{
\item{Adf}{A data frame: Bulk RNA-seq gene expression values of each region, should have same gene names(row names) as in scRNA-seq}
}
\value{
A data frame with all columns are numeric
}
\description{
Check Bulk RNA-seq Region data file
}
\examples{
# Not Run
# data(RNAseq)
# bulkRNA <- check_RNAseq_df(RNAseq$EC)
}
