% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FlyGeneMeta}
\alias{FlyGeneMeta}
\title{Fly gene id and name correspondence}
\format{
data frame, 17558 rows and 2 variables:
\describe{
\item{gene_id}{FBgn id}
\item{gene_name}{Gene symbol}
}
}
\source{
\url{https://flybase.org/}
}
\usage{
FlyGeneMeta
}
\description{
Fly gene id - name transform
}
\keyword{datasets}
