% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RNAseq}
\alias{RNAseq}
\title{RNA-seq RPKM value of four fly gut cell types from five gut regions}
\format{
list, with 4 data frames:
\enumerate{
\item \strong{ISC} A data frame with 15682 rows and 5 columns
\describe{
\item{R1}{Gut Region 1: numeric}
\item{R2}{Gut Region 2: numeric}
\item{R3}{Gut Region 3: numeric}
\item{R4}{Gut Region 4: numeric}
\item{R5}{Gut Region 5: numeric}
}
\item \strong{EB} A data frame with 15682 rows and 5 columns
\describe{
\item{R1}{Gut Region 1: numeric}
\item{R2}{Gut Region 2: numeric}
\item{R3}{Gut Region 3: numeric}
\item{R4}{Gut Region 4: numeric}
\item{R5}{Gut Region 5: numeric}
}
\item \strong{EC} A data frame with 15682 rows and 5 columns
\describe{
\item{R1}{Gut Region 1: numeric}
\item{R2}{Gut Region 2: numeric}
\item{R3}{Gut Region 3: numeric}
\item{R4}{Gut Region 4: numeric}
\item{R5}{Gut Region 5: numeric}
}
\item \strong{EE} A data frame with 15682 rows and 5 columns
\describe{
\item{R1}{Gut Region 1: numeric}
\item{R2}{Gut Region 2: numeric}
\item{R3}{Gut Region 3: numeric}
\item{R4}{Gut Region 4: numeric}
\item{R5}{Gut Region 5: numeric}
}
}
}
\source{
\url{http://flygutseq.buchonlab.com/resources}
}
\usage{
RNAseq
}
\description{
This data set is a list with 4 data.frames, containing the RNA-seq RPKM value of four fly gut cell types(ISC/EB/EC/EE)
from five gut regions, which has a common column in all data.frames.
}
\keyword{datasets}
