% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scRNA}
\alias{scRNA}
\title{Fly gut scRNA-seq by 10xGenomics}
\format{
Seurat object, 2979 cells and 16960 features.
}
\source{
\url{https://www.pnas.org/doi/abs/10.1073/pnas.1916820117}
}
\usage{
scRNA
}
\description{
Hung R J, Hu Y, Kirchner R, et al. A cell atlas of the adult Drosophila midgut. Proceedings of the National Academy
of Sciences, 2020, 117(3): 1514-1523.
}
\keyword{datasets}
