% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNAseq_ScoreRegion.R
\name{scRNAseq_Score_Region}
\alias{scRNAseq_Score_Region}
\title{Calculate Region preference for each cluster}
\usage{
scRNAseq_Score_Region(
  SeuratObj,
  BulkRNAseq.expr,
  UMI.gradient = c(10, 20, 30, 40, 50, 100, 200, 500, 1000, 1500, 2000),
  Genes.gradient = c(10, 20, 30, 40, 50, 100, 200, 300, 400, 500),
  minimal.RNAseq.value = 3.5,
  scRNAseq.expression.cut = 0.5,
  Output.lost.genes = FALSE,
  Customized.GeneList = NULL
)
}
\arguments{
\item{SeuratObj}{Seurat object}

\item{BulkRNAseq.expr}{A data frame: Bulk RNA-seq gene expression values of each region, should have same gene names(row names) as in scRNA-seq}

\item{UMI.gradient}{A numeric vector: UMI cut-off for selecting genes with enough summed expression}

\item{Genes.gradient}{A numeric vector: selected genes numbers for all regions.}

\item{minimal.RNAseq.value}{A numeric: remove genes with low expression value in bulk RNA-seq}

\item{scRNAseq.expression.cut}{A numeric: cut off of the scaled value to define whether a gene is expressed or not.}

\item{Output.lost.genes}{output the genes exist in RNA-seq data while not exist in scRNA-seq data, Default FALSE.}

\item{Customized.GeneList}{A list: customized regional markers}
}
\value{
A list: 1. selected genes for each region; 2. cluster * Region preference for each UMI And Top N genes combination
}
\description{
Calculate Region preference for each cluster
}
\examples{
data(scRNA)
data(RNAseq)
bulkRNAseq <- scRNAseq_Score_Region_Check(RNAseq$EC, FlyGeneMeta)
score.list <- scRNAseq_Score_Region(scRNA, bulkRNAseq)
gene.list <- scRNAseq_Score_Region_ExtractFeatures(score.list)
GeneList <- list()
for(i in colnames(gene.list)){ GeneList[[i]] <- gene.list[,i]}
score.list.customized <- scRNAseq_Score_Region(SeuratObj = scRNA, Customized.GeneList = GeneList)
}
