% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNAseq_ScoreRegion.R
\name{scRNAseq_Score_Region2}
\alias{scRNAseq_Score_Region2}
\title{Predict Region preference for each cluster By gene expression correlation}
\usage{
scRNAseq_Score_Region2(
  SeuratObj,
  BulkRNAseq.expr,
  Method = c("spearman", "pearson", "kendall"),
  Genes.Selection = c("Simple", "Top"),
  Top.foldchange = 3.5,
  Top.UMI.Cutoff = 1,
  Top.numbers = 300,
  Output.lost.genes = FALSE
)
}
\arguments{
\item{SeuratObj}{Seurat object}

\item{BulkRNAseq.expr}{A data frame: Bulk RNA-seq gene expression values of each region, should have same gene names(row names) as in scRNA-seq}

\item{Method}{method used in cor function, one of "spearman", "pearson", "kendall".}

\item{Genes.Selection}{use all genes or only selected genes}

\item{Top.foldchange}{if use selected genes, the cut off of fold change for each Region compared with other regions}

\item{Top.UMI.Cutoff}{if use selected genes, the cut off of UMI for each Region}

\item{Top.numbers}{if use selected genes, the cut off of top gene numbers for each Region}

\item{Output.lost.genes}{output the genes exist in RNA-seq data while not exist in scRNA-seq data, Default FALSE.}
}
\value{
A Region * Cluster matrix
}
\description{
Predict Region preference for each cluster By gene expression correlation
}
\examples{
data(scRNA)
data(RNAseq)
bulkRNAseq <- scRNAseq_Score_Region_Check(RNAseq$EC, FlyGeneMeta)
score.matrix <- scRNAseq_Score_Region2(scRNA, bulkRNAseq, Method = "spearman")
pheatmap::pheatmap(score.matrix)
score.matrix <- scRNAseq_Score_Region2(scRNA, bulkRNAseq, Method = "spearman",
                                       Genes.Selection = "Top")
pheatmap::pheatmap(score.matrix)
}
