% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNAseq_ScoreRegion.R
\name{scRNAseq_Score_Region_evaluate}
\alias{scRNAseq_Score_Region_evaluate}
\title{evaluate the UMI And top n genes combination by Gini index}
\usage{
scRNAseq_Score_Region_evaluate(ScoreList, ...)
}
\arguments{
\item{ScoreList}{A list: Output from scRNAseq_Score_Region function}

\item{...}{Other parameters passed to pheatmap() function}
}
\value{
a Gini index heatmap of all combinations (x: genes used; y: UMI cutoff)
}
\description{
To find a combination with a relatively high Gini index
}
\examples{
data(scRNA)
data(RNAseq)
bulkRNAseq <- scRNAseq_Score_Region_Check(RNAseq$EC, FlyGeneMeta)
score.list <- scRNAseq_Score_Region(scRNA, bulkRNAseq)
scRNAseq_Score_Region_evaluate(score.list)
}
