% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNAseq_ScoreRegion.R
\name{scRNAseq_Score_Region_ExtractFeatures}
\alias{scRNAseq_Score_Region_ExtractFeatures}
\title{Extract features}
\usage{
scRNAseq_Score_Region_ExtractFeatures(ScoreList, UMI = NULL, TopGene = NULL)
}
\arguments{
\item{ScoreList}{A list: Output from scRNAseq_Score_Region function}

\item{UMI}{UMI cutoff defined in scRNAseq_Score_Region function - UMI.gradient}

\item{TopGene}{Top n genes defined in scRNAseq_Score_Region function - Genes.gradient}
}
\value{
\enumerate{
\item A venn diagram of the genes from different regions; 2.A data frame of the extracted genes for each region.
}
}
\description{
Extract features
}
\examples{
data(scRNA)
data(RNAseq)
bulkRNAseq <- scRNAseq_Score_Region_Check(RNAseq$EC, FlyGeneMeta)
score.list <- scRNAseq_Score_Region(scRNA, bulkRNAseq)
## futile.logger::flog.threshold(futile.logger::ERROR, name = "VennDiagramLogger")
scRNAseq_Score_Region_ExtractFeatures(score.list)
scRNAseq_Score_Region_ExtractFeatures(score.list, 100, 100)
}
